+++ src/collective/taxonomy/behavior.py
@@ -9,7 +9,6 @@ from plone.autoform.interfaces import (
     IFormFieldProvider
 )
 
-from plone.app.dexterity import PloneMessageFactory as _pmf
 from plone.behavior.interfaces import IBehavior
 from plone.dexterity.interfaces import IDexterityContent
 from plone.indexer.interfaces import IIndexer
@@ -30,8 +29,10 @@ from zope.component.hooks import getSite
 from zope.interface import implements, alsoProvides
 from zope.component import getUtility
 
-from .i18n import MessageFactory as _
-from .indexer import TaxonomyIndexer
+from collective.taxonomy.i18n import CollectiveTaxonomyMessageFactory as _
+from collective.taxonomy.i18n import _pmf
+from collective.taxonomy.indexer import TaxonomyIndexer
+
 
 logger = logging.getLogger("collective.taxonomy")