+++ src/collective/taxonomy/testing.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
@@ -31,3 +32,7 @@ class Fixture(PloneSandboxLayer):
 FIXTURE = Fixture()
 INTEGRATION_TESTING = IntegrationTesting(
     bases=(FIXTURE,), name="TaxonomyFixture:Integration")
+
+FUNCTIONAL_TESTING = FunctionalTesting(
+    bases=(FIXTURE,), name="TaxonomyFixture:Functional")
+